管理者チェックを追加

+++ src/main/java/actions/EmployeeAction.java
@@ -34,8 +34,10 @@ public class EmployeeAction extends ActionBase {
      * @throws ServletException
      * @throws IOException
      */
-    public void index() throws ServletException,IOException{
+    public void index() throws ServletException, IOException {
 
+        //管理者かどうかのチェック
+        if (checkAdmin()) { //チャプター8.6追記
         //指定されたページ数の一覧画面に表示するデータを取得
         int page = getPage();
         List<EmployeeView> employees = service.getPerPage(page);
@@ -57,6 +59,8 @@ public class EmployeeAction extends ActionBase {
         }
         //一覧画面を表示
         forward(ForwardConst.FW_EMP_INDEX);
+        }
+
 
     }
     /*
@@ -65,12 +69,15 @@ public class EmployeeAction extends ActionBase {
      * @throws IOException
      */
     public void entryNew() throws ServletException,IOException{
+        //管理者かどうかのチェック
+        if (checkAdmin()) { //チャプター8.6追記
 
         putRequestScope(AttributeConst.TOKEN,getTokenId());//CSRF対策
         putRequestScope(AttributeConst.EMPLOYEE,new EmployeeView());
 
         //新規登録画面を表示
         forward(ForwardConst.FW_EMP_NEW);
+        }
     }
 
     /*
@@ -81,7 +88,7 @@ public class EmployeeAction extends ActionBase {
     public void create() throws ServletException,IOException{
 
         //CSRF対策
-        if(checkToken()) {
+         if (checkAdmin() && checkToken()) { //チャプター8.6追記
 
             //パラメータの値を元に従業員情報のインスタンスを作成する
             EmployeeView ev = new EmployeeView(
@@ -124,12 +131,15 @@ public class EmployeeAction extends ActionBase {
 
 
     }
-    /**
+    /*
      * 詳細画面を表示する
      * @throws ServletException
      * @throws IOException
      */
     public void show() throws ServletException, IOException {
+        //管理者かどうかのチェック
+        if (checkAdmin()) { //チャプター8.6追記
+    }
 
         //idを条件に従業員データを取得する
         EmployeeView ev = service.findOne(toNumber(getRequestParam(AttributeConst.EMP_ID)));
@@ -153,6 +163,9 @@ public class EmployeeAction extends ActionBase {
      * @throws IOException
      */
     public void edit() throws ServletException,IOException{
+        //管理者かどうかのチェック
+        if (checkAdmin()) { //チャプター8.6追記
+
         //idを条件に従業員データを取得する
         EmployeeView ev = service.findOne(toNumber(getRequestParam(AttributeConst.EMP_ID)));
          if (ev == null || ev.getDeleteFlag() == AttributeConst.DEL_FLAG_TRUE.getIntegerValue()) {
@@ -166,6 +179,7 @@ public class EmployeeAction extends ActionBase {
 
          //編集画面を表示する
          forward(ForwardConst.FW_EMP_EDIT);
+        }
     }
     /*
      * 更新を行う
@@ -174,7 +188,7 @@ public class EmployeeAction extends ActionBase {
      */
     public void update() throws ServletException,IOException{
         //CSRF対策tokenのチェック
-        if(checkToken()) {
+        if (checkAdmin() && checkToken()) { //チャプター8.6追記
             //パラメータの値を元に従業員のインスタンスを作成する
             EmployeeView ev = new EmployeeView(
                     toNumber(getRequestParam(AttributeConst.EMP_ID)),
@@ -219,7 +233,7 @@ public class EmployeeAction extends ActionBase {
      */
     public void destroy() throws ServletException,IOException{
         //CSRF対策
-        if(checkToken()) {
+        if (checkAdmin() && checkToken()) { //チャプター8.6追記
             //idを条件に従業員データを論理削除する
             service.destroy(toNumber(getRequestParam(AttributeConst.EMP_ID)));
 
@@ -228,6 +242,24 @@ public class EmployeeAction extends ActionBase {
 
             //一覧画面にリダイレクト
             redirect(ForwardConst.ACT_EMP, ForwardConst.CMD_INDEX);
+        }
+    }
+
+    /*
+     * ログイン中の従業員が管理者かどうかチェックし、管理者でなければエラー画面を表示
+     * true: 管理者 false: 管理者ではない
+     * @throws ServletException
+     * @throws IOException
+     */
+    private boolean checkAdmin() throws ServletException,IOException {
+        //セッションからログイン中の従業員情報を取得
+        EmployeeView ev = (EmployeeView)getSessionScope(AttributeConst.LOGIN_EMP);
+        //管理者でなければエラー画面を表示
+        if(ev.getAdminFlag() != AttributeConst.ROLE_ADMIN.getIntegerValue()) {
+            forward(ForwardConst.FW_ERR_UNKNOWN);
+            return false;
+        }else {
+            return true;
         }
     }
 }
